Moves newUpgradeJujuCommandForTest into the test module, which is the only place that it is used.
